+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.0.2 (next release)
+
+- removed support for deprecated `steadybit.com/extension-auto-discovery` annotations
+- revisited debug-logging
+
 ## v0.0.1
 
 - Initial release